sort symmetric ops
This should both allow more dedup (though it doesn't hit anything in
prospero) and should allow more rhs-in-reg, reducing spills, which it
does quite a lot in prospero, for things like this:

    ...
    _23 neg _22
    _24 max _20 _23
    _25 max _24 _19
    _26 max _25 _1d
    ...

_25 gets sorted to max _19 _24, and _26 is sorted to max _1d _25.
Notice how this can never add spills: i-1 is the maximum possible
value for rhs, so if we weren't spilling before, sort() does nothing.

So this should be unambiguously good.  However... prospero runs a little
slower.  I'm guessing this is because the spill and non-spill op
variants get separate codegen and branch prediction points, and that's
muddled here when more goes through the non-spills?  Keeping it because
it's interesting and reduces scratch memory usage, but a puzzler.